+++ packages/stream/stream_server.js
@@ -7,7 +7,7 @@ Meteor._routePolicy.declare('/sockjs/', 'network');
 // custom serverId which you only change when something worth pushing
 // to clients immediately happens.
 __meteor_runtime_config__.serverId =
-  process.env.SERVER_ID ? process.env.SERVER_ID : Meteor.uuid();
+  process.env.SERVER_ID ? process.env.SERVER_ID : Random.id();
 
 Meteor._StreamServer = function () {
   var self = this;