+++ packages/tinytest/tinytest_client.js
@@ -2,7 +2,7 @@
 // the server. Sets a 'server' flag on test results that came from the
 // server.
 Meteor._runTestsEverywhere = function (onReport, onComplete) {
-  var runId = LocalCollection.uuid();
+  var runId = Random.id();
   var localComplete = false;
   var remoteComplete = false;
   var done = false;
@@ -25,14 +25,16 @@ Meteor._runTestsEverywhere = function (onReport, onComplete) {
       // we really only should see one runId here.
       if (msg.id !== runId)
         return;
-      _.each(msg.set, function (report) {
+      // This will only work for added & changed messages.
+      // hope that is all you get.
+      _.each(msg.fields, function (report) {
         _.each(report.events, function (event) {
           delete event.cookie; // can't debug a server test on the client..
         });
         report.server = true;
         onReport(report);
       });
-    }
+    }  
   });
 
   var handle = Meteor.subscribe(Meteor._ServerTestResultsSubscription, runId);